RefactoCNN-System/requirements.txt
torch>=2.0
numpy>=1.24
pyyaml>=6.0
tqdm>=4.66
scikit-learn>=1.3
javalang>=0.13.0
tensorboard>=2.14
optuna>=3.6
flask>=3.0
